Silence warning with gcc 15 and O0/01
When compiling with low optimisation level gcc complains

   route.c:1783:44: error: conversion from ‘int’ to ‘uint8_t’ {aka ‘unsigned char’} may change value [-Werror=conversion]
    1783 |             r6->network.s6_addr[byte--] &= (0xff << bits_to_clear);

With higher it apparently detects that this is cannot be and does not
trigger the warning. Explicitly cast to uint8_t to silence this warning

Acked-by: Frank Lichtenheld <frank@lichtenheld.com>
Gerrit URL: https://gerrit.openvpn.net/c/openvpn/+/1731
Message-Id: <20260723180445.17392-1-gert@greenie.muc.de>
URL: https://www.mail-archive.com/openvpn-devel@lists.sourceforge.net/msg37816.html

+++ src/openvpn/route.c
@@ -1780,7 +1780,7 @@ route_ipv6_clear_host_bits(struct route_ipv6 *r6)
         }
         else
         {
-            r6->network.s6_addr[byte--] &= (0xff << bits_to_clear);
+            r6->network.s6_addr[byte--] &= (uint8_t)(0xff << bits_to_clear);
             bits_to_clear = 0;
         }
     }
